ml-flask/requirements.txt
flask
torch
torchvision
torchaudio
transformers
accelerate
bitsandbytes
Pillow
numpy
ultralytics
python-dotenv
requests
sentencepiece